refactor(adapters): relocate registry client into adapters/driven/registry (hexagonal P4 sweep)
Pure git mv of regis/registry/{auth,client,parser}.py →
regis/adapters/driven/registry/ (joining the regctl/registry ImageInspector
adapters) + import-path updates across the ~15 importers (composition, commands,
utils/regctl, the inspector adapters, tests + their patch targets). Vestigial
regis/registry/__init__.py removed. RegistryClient still subclasses the core
RegistryError (adapter→core, allowed). No behaviour change. Third subtree of
the P4 sweep.
