+++ regis/adapters/driven/registry/auth.py
