+++ regis/adapters/driven/registry/client.py
