+++ regis/adapters/driven/registry/regctl_image_inspector.py
@@ -8,9 +8,9 @@ from collections.abc import Iterator
 from contextlib import contextmanager
 from typing import Any
 
+from regis.adapters.driven.registry.client import RegistryClient
 from regis.core.domain.errors import RegistryError
 from regis.core.ports.image_inspector import ImageInspector
-from regis.registry.client import RegistryClient
 from regis.utils.regctl import image_ref, run_regctl
