--- regis/registry/__init__.py
@@ -1,1 +1,0 @@
-"""Docker Registry V2 API client."""
